[refactor]: 재사용 가능성이 있는 컴포넌트 분리 (#1)
