+++ src/components/Header.tsx
@@ -1,0 +1,12 @@
+import Link from 'next/link';
+
+const Header = () => {
+  return (
+    <header className="flex justify-between mt-5">
+      <h1 className="font-bold text-lg">Share The Journey</h1>
+      <Link href="/login">Login</Link>
+    </header>
+  );
+};
+
+export default Header;
